Complete architectural restructure
