+++ src/hooks/useRoomPolling.js
@@ -1,5 +1,5 @@
 import { useEffect, useRef } from "react";
-import { getRoom } from "../logic/gameRoomApi";
+import { getRoom } from "../utils/gameRoomApi";
 
 export function useRoomPolling({ roomId, enabled, onUpdate, onAbandoned }) {
   const currentPokemonIdRef = useRef(null);
@@ -19,15 +19,7 @@ export function useRoomPolling({ roomId, enabled, onUpdate, onAbandoned }) {
 
         onUpdate(gameState);
 
-        if (
-          gameState.pokemon &&
-          gameState.pokemon.id !== currentPokemonIdRef.current
-        ) {
-          currentPokemonIdRef.current = gameState.pokemon.id;
-          console.log(
-            `🎮 Pokemon: ${gameState.pokemon.name} (ID: ${gameState.pokemon.id})`,
-          );
-        }
+
       } catch {}
     }, 1000);
