--- src/logic/avatars.js
@@ -1,15 +1,0 @@
-import wizardAvatar from "../assets/avatars/wizard.png";
-import knightAvatar from "../assets/avatars/knight.png";
-import ninjaAvatar from "../assets/avatars/ninja.png";
-import sorceressAvatar from "../assets/avatars/sorceress.png";
-import valkyrieAvatar from "../assets/avatars/valkyrie.png";
-import assassinAvatar from "../assets/avatars/assassin.png";
-
-export const avatars = [
-  { key: `wizard`, image: wizardAvatar },
-  { key: `knight`, image: knightAvatar },
-  { key: `ninja`, image: ninjaAvatar },
-  { key: `sorceress`, image: sorceressAvatar },
-  { key: `valkyrie`, image: valkyrieAvatar },
-  { key: `assassin`, image: assassinAvatar },
-];
