+++ src/pages/HomePage.jsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import { useState } from "react";
 import { games } from "../data/gamesData";
-import { avatars } from "../data/avatars";
+import { avatars } from "../utils/avatars";
 import "../styles/HomePage.css";
 
 const LEFT_WIDTH = 260;
