+++ src/pages/pokemon/components/HiddenPicturePreview.jsx
