+++ src/pages/pokemon/components/ScoreBar.jsx
