+++ src/pages/pokemon/logic/pokemonLogic.js
@@ -146,14 +146,3 @@ export function playerLeft(state, playerId) {
 
   return { ...state, players, currentPlayerIndex: nextIndex };
 }
-
-export async function fetchRandomPokemon() {
-  const id = Math.floor(Math.random() * 151) + 1;
-  const res = await fetch(`https://pokeapi.co/api/v2/pokemon/${id}`);
-  const data = await res.json();
-  return {
-    id: data.id,
-    name: data.name,
-    image: data.sprites.other["official-artwork"].front_default,
-  };
-}
