+++ src/pages/pokemon/utils/pokemonApi.js
@@ -1,0 +1,10 @@
+export async function fetchRandomPokemon() {
+  const id = Math.floor(Math.random() * 151) + 1;
+  const res = await fetch(`https://pokeapi.co/api/v2/pokemon/${id}`);
+  const data = await res.json();
+  return {
+    id: data.id,
+    name: data.name,
+    image: data.sprites.other["official-artwork"].front_default,
+  };
+}
