+++ src/pages/rock-paper-scissors/components/GameHeader.jsx
