+++ src/pages/rock-paper-scissors/components/HighScoresSection.jsx
