+++ src/pages/rock-paper-scissors/components/HistorySection.jsx
