+++ src/pages/rock-paper-scissors/components/ScoreBoard.jsx
