+++ src/utils/settings.js
